Remove location argument from constructor for Downloader class
Having the `location` argument be part of the constructor for the `Downloader` class means that instances may only be used once, since it forces all downloaded files to be in the same location on disk.

If the `location` argument is removed and then supplied as an argument to each instance method, then the same instance may be used multiple times, with each set of downloaded files going to [potentially -- as determined by the caller] a different location on disk.